Add ACM application files for near-edge cluster set
The changes in this commit are equivalent to the changes that @LaVLaS
made in #37, with some minor changes (that I also added as comment
suggestions to on his PR).

To get this to sync correctly, I had to modify the existing
`poc-near-edge` cluster set in ACM to add a namespace binding for the
`rhoai-edge-acm` namespace, where this subscription is created. I'm
not sure how that change could be encoded here in the repo.