Use ACM app per app to ensure correct NS placement
It seems like all resources get created in a namespace on the target
clusters with the same name as the namespace that the subscription on
the hub cluster is created in, despite the resources explicitly having
a different namespace set.

This change adds new subscriptions such that the two model apps get
created in different namespaces on the target clusters.