Learn how to push image to many tags
JIRA: https://issues.redhat.com/browse/RHOAIENG-2758

Before this change, users could specify one image tag reference to
push the built and tested image to, and we'd have encouraged always
having a unique part of the tag, such as by appending the pipelinerun
id, as in our examples. While this worked well, it didn't make it easy
to determine from looking at an image registry, which of the unique
tags is the current one at any given time.

This change allows for both uniqueness and meaning. The simplest
example scenario is where there's a primary tag that's unique, which
has the model version and pipelinerun id; then a secondary tag with
just the model version, which tracks the current blessed image for
that model version. Quay (and possibly other registries) even has a UI
feature for visualising that different tags actually point to the same
manifest.