Remove need to explicitly create imagestream
JIRA: https://issues.redhat.com/browse/RHOAIENG-2758

Before this change, we were explicitly creating the imagestream for
the candidate image from the build task, just so that we could set the
value of `spec.lookupPolicy.local` to `true`. While we could have left
the imagestream to be implicitly created with default config just by
pushing an image with the equivalent name, this value would not have had
the desired value.

When the value of `spec.lookupPolicy.local` is at the default of
false, then referring to an image by just `<name>:<tag>` without fully
qualifying it with the registry domain, then it would default to the
`docker.io` registry, and ignore the repository with that name in the
internal registry in the cluster.

By setting the container image path to the fully qualified name,
including the registry part (the same has how we specify it with in
the parameter to the buildah task), then we don't need to care about
the `spec.lookupPolicy.local` field, and can rely on the default
creation of the ImageStream.