Add alternative GitOps flow without OCM/ACM
JIRA: https://issues.redhat.com/browse/RHOAIENG-2787

We love OCM/ACM and recommend using it where possible, but if you've
got your own multi-cluster management solution and don't want to have
to replace it to use the solutions in this repo, then you can instead
use this as a starting point to integrate it yourself.

The steps are intended to be documented in such a way that they could
be either executed manually or adapted to fit into any system.