Remove imageReferenceFilePath from gitops-update-pipeline
JIRA: https://issues.redhat.com/browse/RHOAIENG-3128

This parameter was limiting, as it restricted the pipeline to update
only one file, even though a user may wish to update the image
reference in multiple files in the repo.

With the new `yq-script` parameter, this `imageReferenceFilePath`
parameter is no longer needed, as the user can just specify the file
path(s) that they wish to update with each yq command directly in the
script, rather than referring to the value of this other param.