Show example of uber-pipeline containing everything
JIRA: https://issues.redhat.com/browse/RHOAIENG-4982

This shows what it would be like to have the aiedge-e2e pipeline also
include all of the tasks from the gitops update pipeline.

There are advantages and disadvantages to this approach:

Advantages:

- Compared to the EventListener/Trigger approach, no Tekton
EventListener pod needs to be running all the time.

Disadvantages:
- The aiedge-e2e pipeline becomes significantly more complex, and it
now needs to know how to do everything that the gitops update pipeline
knows how to do.
- If we change anything in the gitops update pipeline, we need to
remember to also update that in this pipeline.
- Making it optional (for example, for users who don't us a forge that
implements the GitHub PR API) would require a 10th additional parameter